@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 
 class DisciplinaComponents extends StatelessWidget {
-  final titolo;
-  final horas;
+  final String titolo;
+  final String horas;
   final Color horasC;
   final Color titoloC;
-  DisciplinaComponents(
-      {Key key, this.titolo, this.horas, this.horasC, this.titoloC})
-      : super(key: key);
+
+  const DisciplinaComponents({Key? key, required this.titolo, required this.horas, required this.horasC, required this.titoloC}) : super(key: key);
+  
   @override
   Widget build(BuildContext context) {
     return Column(
@@ -17,22 +17,22 @@ class DisciplinaComponents extends StatelessWidget {
           width: 45,
           height: 30,
           child: Text(
-            this.titolo,
-            style: TextStyle(
+            titolo,
+            style: const TextStyle(
               color: Colors.black,
               fontWeight: FontWeight.bold,
               fontSize: 20,
             ),
           ),
           decoration: BoxDecoration(
-            color: this.titoloC,
+            color: titoloC,
           ),
         ),
         Container(
           alignment: Alignment.center,
           child: Text(
-            this.horas,
-            style: TextStyle(
+            horas,
+            style: const TextStyle(
               color: Colors.white,
               fontWeight: FontWeight.bold,
               fontSize: 12,
@@ -41,8 +41,8 @@ class DisciplinaComponents extends StatelessWidget {
           width: 45,
           height: 20,
           decoration: BoxDecoration(
-              color: this.horasC,
-              borderRadius: BorderRadius.only(
+              color: horasC,
+              borderRadius: const BorderRadius.only(
                 bottomLeft: Radius.circular(5.0),
                 bottomRight: Radius.circular(5.0),
               )),
