@@ -6,12 +6,12 @@ class MyMenu extends StatelessWidget {
   final MaterialColor warn;
   final Function navi;
 
-  const MyMenu({Key key, this.title, this.icon, this.warn, this.navi})
+  const MyMenu({Key? key, required   this.title,required this.icon, required this.warn, required this.navi})
       : super(key: key);
   @override
   Widget build(BuildContext context) {
     return Card(
-      margin: EdgeInsets.all(8),
+      margin: const EdgeInsets.all(8),
       child: InkWell(
         onTap: () {
           Navigator.of(context).pushNamed('pagamento');
@@ -22,7 +22,7 @@ class MyMenu extends StatelessWidget {
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
             crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
+            children: const [
               Icon(Icons.home),
               Text("home"),
             ],
