@@ -6,6 +6,8 @@ import 'package:sistema_escolar/res/colors/app_colors.dart';
 import 'package:sistema_escolar/utils/rotas_util.dart';
 
 class DashboardPage extends StatefulWidget {
+  const DashboardPage({Key? key}) : super(key: key);
+
   @override
   _DashboardPageState createState() => _DashboardPageState();
 }
@@ -17,109 +19,109 @@ class _DashboardPageState extends State<DashboardPage> {
       appBar: AppBar(
         backgroundColor: AppColors.green,
         elevation: 0,
-        title: Text('DashBoard'),
+        title: const Text('DashBoard'),
       ),
       drawer: drawerComponts(),
       body: SingleChildScrollView(
         child: Column(
           children: [
-            MiniProfileComponts(),
+            const MiniProfileComponts(),
+
+            //Container
             SizedBox(
               height: 400,
-              child: Container(
-                child: GridView.count(
-                  crossAxisCount: 4,
-                  children: [
-                    Card(
-                      margin: EdgeInsets.all(8),
-                      child: InkWell(
-                        onTap: () {
-                          nevegarTela(context, 'falta');
-                        },
-                        splashColor: Colors.green,
-                        child: Container(
-                          alignment: Alignment.center,
-                          child: Column(
-                            mainAxisAlignment: MainAxisAlignment.center,
-                            crossAxisAlignment: CrossAxisAlignment.center,
-                            children: [
-                              Icon(
-                                Icons.feedback,
-                              ),
-                              Text("Falta"),
-                            ],
-                          ),
+              child: GridView.count(
+                crossAxisCount: 4,
+                children: [
+                  Card(
+                    margin: const EdgeInsets.all(8),
+                    child: InkWell(
+                      onTap: () {
+                        nevegarTela(context, 'falta');
+                      },
+                      splashColor: Colors.green,
+                      child: Container(
+                        alignment: Alignment.center,
+                        child: Column(
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          crossAxisAlignment: CrossAxisAlignment.center,
+                          children: const [
+                            Icon(
+                              Icons.feedback,
+                            ),
+                            Text("Falta"),
+                          ],
                         ),
                       ),
                     ),
-                    Card(
-                      margin: EdgeInsets.all(8),
-                      child: InkWell(
-                        onTap: () {
-                          Navigator.of(context).pushNamed('pagamentos');
-                        },
-                        splashColor: Colors.green,
-                        child: Container(
-                          alignment: Alignment.center,
-                          child: Column(
-                            mainAxisAlignment: MainAxisAlignment.center,
-                            crossAxisAlignment: CrossAxisAlignment.center,
-                            children: [
-                              Icon(Icons.payment),
-                              Text("Mensalidades"),
-                            ],
-                          ),
+                  ),
+                  Card(
+                    margin: const EdgeInsets.all(8),
+                    child: InkWell(
+                      onTap: () {
+                        Navigator.of(context).pushNamed('pagamentos');
+                      },
+                      splashColor: Colors.green,
+                      child: Container(
+                        alignment: Alignment.center,
+                        child: Column(
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          crossAxisAlignment: CrossAxisAlignment.center,
+                          children: const [
+                            Icon(Icons.payment),
+                            Text("Mensalidades"),
+                          ],
                         ),
                       ),
                     ),
-                    Card(
-                      margin: EdgeInsets.all(8),
-                      child: InkWell(
-                        onTap: () {
-                          Navigator.of(context).pushNamed('horario');
-                        },
-                        splashColor: Colors.green,
-                        child: Container(
-                          alignment: Alignment.center,
-                          child: Column(
-                            mainAxisAlignment: MainAxisAlignment.center,
-                            crossAxisAlignment: CrossAxisAlignment.center,
-                            children: [
-                              Icon(Icons.timer_off),
-                              Text("Horario"),
-                            ],
-                          ),
+                  ),
+                  Card(
+                    margin: const EdgeInsets.all(8),
+                    child: InkWell(
+                      onTap: () {
+                        Navigator.of(context).pushNamed('horario');
+                      },
+                      splashColor: Colors.green,
+                      child: Container(
+                        alignment: Alignment.center,
+                        child: Column(
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          crossAxisAlignment: CrossAxisAlignment.center,
+                          children: const [
+                            Icon(Icons.timer_off),
+                            Text("Horario"),
+                          ],
                         ),
                       ),
                     ),
-                    Card(
-                      margin: EdgeInsets.all(8),
-                      child: InkWell(
-                        onTap: () {
-                          Navigator.of(context).pushNamed('notas');
-                        },
-                        splashColor: Colors.green,
-                        child: Container(
-                          alignment: Alignment.center,
-                          child: Column(
-                            mainAxisAlignment: MainAxisAlignment.center,
-                            crossAxisAlignment: CrossAxisAlignment.center,
-                            children: [
-                              Icon(Icons.timer_off),
-                              Text("Notas"),
-                            ],
-                          ),
+                  ),
+                  Card(
+                    margin: const EdgeInsets.all(8),
+                    child: InkWell(
+                      onTap: () {
+                        Navigator.of(context).pushNamed('notas');
+                      },
+                      splashColor: Colors.green,
+                      child: Container(
+                        alignment: Alignment.center,
+                        child: Column(
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          crossAxisAlignment: CrossAxisAlignment.center,
+                          children: const [
+                            Icon(Icons.timer_off),
+                            Text("Notas"),
+                          ],
                         ),
                       ),
                     ),
-                  ],
-                ),
+                  ),
+                ],
               ),
             )
           ],
         ),
       ),
-      bottomNavigationBar: ButtomAppBarComponents(),
+      bottomNavigationBar: const ButtomAppBarComponents(),
     );
   }
 }
