@@ -5,6 +5,8 @@ import 'package:sistema_escolar/components/mini_profile_componts.dart';
 import 'package:sistema_escolar/res/colors/app_colors.dart';
 
 class HorarioPage extends StatefulWidget {
+  const HorarioPage({Key? key}) : super(key: key);
+
   @override
   _HorarioPageState createState() => _HorarioPageState();
 }
@@ -17,14 +19,14 @@ class _HorarioPageState extends State<HorarioPage> {
         centerTitle: true,
         elevation: 0,
         backgroundColor: AppColors.green,
-        title: Text(
+        title: const Text(
           'Horário',
           style: TextStyle(
             fontSize: 25,
             fontWeight: FontWeight.bold,
           ),
         ),
-        actions: [
+        actions: const [
           Icon(
             Icons.notifications,
             color: Colors.white,
@@ -43,119 +45,119 @@ class _HorarioPageState extends State<HorarioPage> {
       ),
       body: Column(
         children: [
-          MiniProfileComponts(),
-          SizedBox(height: 20),
-          Expanded(
+         const MiniProfileComponts(),
+         const SizedBox(height: 20),
+         Expanded(
             flex: 2,
             child: PageView.builder(
               itemBuilder: (context, int index) {
                 return Container(
-                  margin: EdgeInsets.all(8),
+                  margin: const EdgeInsets.all(8),
                   child: Column(
                     children: [
                       diasSemana(
                           dia: "Segunda-Feira", ano: "4 de Janeiro de 2021"),
-                      SizedBox(height: 20),
+                      const SizedBox(height: 20),
                       Column(
                         mainAxisAlignment: MainAxisAlignment.center,
                         crossAxisAlignment: CrossAxisAlignment.center,
                         children: [
-                          Text(
+                          const Text(
                             "MANHA",
                             style: TextStyle(
                                 color: Colors.black,
                                 fontWeight: FontWeight.bold,
                                 fontSize: 20),
                           ),
-                          SizedBox(height: 20),
+                          const SizedBox(height: 20),
                           Row(
                             mainAxisAlignment: MainAxisAlignment.center,
                             crossAxisAlignment: CrossAxisAlignment.center,
-                            children: [
+                            children: const [
                               DisciplinaComponents(
                                   horas: '07:15',
                                   titolo: 'GEO',
-                                  titoloC: Colors.brown[100],
+                                  titoloC: Colors.brown,
                                   horasC: Colors.pink),
                               SizedBox(width: 5),
                               DisciplinaComponents(
                                   horas: '08:15',
                                   titolo: 'GEO',
-                                  titoloC: Colors.brown[100],
+                                  titoloC: Colors.brown,
                                   horasC: Colors.pink),
                               SizedBox(width: 5),
                               DisciplinaComponents(
                                   horas: '08:15',
                                   titolo: 'FIS',
-                                  titoloC: Colors.brown[100],
+                                  titoloC: Colors.brown,
                                   horasC: Colors.pink),
-                              SizedBox(width: 5),
+                             SizedBox(width: 5),
                               DisciplinaComponents(
                                   horas: '08:15',
                                   titolo: 'FIS',
-                                  titoloC: Colors.brown[100],
+                                  titoloC: Colors.brown,
                                   horasC: Colors.pink),
                               SizedBox(width: 5),
                               DisciplinaComponents(
                                   horas: '08:15',
                                   titolo: 'TOP',
-                                  titoloC: Colors.brown[100],
+                                  titoloC: Colors.brown,
                                   horasC: Colors.pink),
-                              SizedBox(width: 5),
+                            SizedBox(width: 5),
                               DisciplinaComponents(
                                   horas: '08:15',
                                   titolo: 'TOP',
-                                  titoloC: Colors.brown[100],
+                                  titoloC: Colors.brown,
                                   horasC: Colors.pink),
                             ],
                           ),
-                          SizedBox(height: 20),
-                          Text(
+                          const SizedBox(height: 20),
+                          const Text(
                             "MANHA",
                             style: TextStyle(
                                 color: Colors.black,
                                 fontWeight: FontWeight.bold,
                                 fontSize: 20),
                           ),
-                          SizedBox(height: 20),
+                          const SizedBox(height: 20),
                           Row(
                             mainAxisAlignment: MainAxisAlignment.center,
                             crossAxisAlignment: CrossAxisAlignment.center,
-                            children: [
+                            children: const [
                               DisciplinaComponents(
                                   horas: '07:15',
                                   titolo: 'FPF',
-                                  titoloC: Colors.brown[100],
+                                  titoloC: Colors.brown,
                                   horasC: Colors.pink),
                               SizedBox(width: 5),
                               DisciplinaComponents(
                                   horas: '08:15',
                                   titolo: 'FPF',
-                                  titoloC: Colors.brown[100],
+                                  titoloC: Colors.brown,
                                   horasC: Colors.pink),
                               SizedBox(width: 5),
                               DisciplinaComponents(
                                   horas: '08:15',
                                   titolo: 'FPF',
-                                  titoloC: Colors.brown[100],
+                                  titoloC: Colors.brown,
                                   horasC: Colors.pink),
                               SizedBox(width: 5),
                               DisciplinaComponents(
                                   horas: '08:15',
                                   titolo: 'EDF',
-                                  titoloC: Colors.brown[100],
+                                  titoloC: Colors.brown,
                                   horasC: Colors.pink),
                               SizedBox(width: 5),
                               DisciplinaComponents(
                                   horas: '08:15',
                                   titolo: 'EDF',
-                                  titoloC: Colors.brown[100],
+                                  titoloC: Colors.brown,
                                   horasC: Colors.pink),
                               SizedBox(width: 5),
                               DisciplinaComponents(
                                   horas: '08:15',
                                   titolo: '',
-                                  titoloC: Colors.brown[100],
+                                  titoloC: Colors.brown,
                                   horasC: Colors.pink),
                             ],
                           ),
