@@ -1,9 +1,9 @@
-
 import 'package:flutter/material.dart';
-import 'package:sistema_escolar/components/alter_dialog_component.dart';
 import 'package:sistema_escolar/components/form_field_components.dart';
 
 class LoginPage extends StatefulWidget {
+  const LoginPage({Key? key}) : super(key: key);
+
   @override
   _LoginPageState createState() => _LoginPageState();
 }
@@ -19,15 +19,15 @@ class _LoginPageState extends State<LoginPage> {
       onPressed: () {
         _cliqueButton(context);
       },
-      child: Text(
+      child: const Text(
         "Entrar",
         style: TextStyle(fontSize: 20),
       ),
-  //    color: Colors.blue,
+      //    color: Colors.blue,
     );
     return Scaffold(
       appBar: AppBar(
-        title: Text(''),
+        title: const Text(''),
       ),
       body: Form(
         key: _formkey,
@@ -41,7 +41,6 @@ class _LoginPageState extends State<LoginPage> {
                 hint: 'Digite o email',
                 label: 'E-mail',
                 controller: _crtEmail,
-                validator: _validaLogin,
                 senha: false,
                 teclado: TextInputType.text,
               ),
@@ -52,7 +51,6 @@ class _LoginPageState extends State<LoginPage> {
                 hint: 'Digite a senha',
                 label: 'Senha',
                 controller: _crtSenha,
-                validator: _validaSenha,
                 senha: true,
                 teclado: TextInputType.text,
               ),
@@ -67,46 +65,11 @@ class _LoginPageState extends State<LoginPage> {
     );
   }
 
-  String _validaLogin(String texto) {
-    if (texto.isEmpty) {
-      return "Digite o Email";
-    }
-    if (!texto.contains("@") || !texto.contains(".")) {
-      return "Email envalido";
-    }
-
-    return null;
-  }
-
-  String _validaSenha(String texto) {
-    if (texto.isEmpty) {
-      return "Digite a Senha";
-    }
-    if (texto.length < 8) {
-      return "A Senha Deve ter mais de 8 Caractere";
-    }
-    return null;
-  }
-
   _cliqueButton(BuildContext context) async {
-    bool formOk = _formkey.currentState.validate();
+    bool formOk = _formkey.currentState!.validate();
     if (!formOk) {
       return;
     }
-    String email = _crtEmail.text;
-    String senha = _crtSenha.text;
-
-    print("login $email $senha");
-    var usuario;
-    if (usuario != null) {
-      print("==> $usuario");
-      _nevegarTela(context);
-    } else {
-      alert(context, "Login Invalido");
-    }
   }
 
-  _nevegarTela(BuildContext context) {
-    Navigator.pushReplacementNamed(context, 'users');
-  }
 }
