@@ -4,6 +4,8 @@ import 'package:sistema_escolar/res/colors/app_colors.dart';
 import 'package:sistema_escolar/utils/rotas_util.dart';
 
 class NotaPage extends StatefulWidget {
+  const NotaPage({Key? key}) : super(key: key);
+
   @override
   _NotaPageState createState() => _NotaPageState();
 }
@@ -17,13 +19,13 @@ class _NotaPageState extends State<NotaPage> {
         backgroundColor: AppColors.green,
         elevation: 0,
         leading: IconButton(
-          icon: Icon(Icons.arrow_back),
+          icon: const Icon(Icons.arrow_back),
           onPressed: () {
             nevegarTela(context, 'dashboard');
           },
         ),
-        title: Text('Notas'),
-        actions: [
+        title: const Text('Notas'),
+        actions: const [
           Icon(
             Icons.notifications,
             color: Colors.white,
@@ -42,7 +44,7 @@ class _NotaPageState extends State<NotaPage> {
       ),
       body: Column(
         children: [
-          MiniProfileComponts(),
+          const MiniProfileComponts(),
           SizedBox(
             height: 400,
             child: SingleChildScrollView(
@@ -53,7 +55,7 @@ class _NotaPageState extends State<NotaPage> {
                   Card(
                     child: Column(
                       children: [
-                        Text(
+                        const Text(
                           "FISICA",
                           style: TextStyle(
                             fontSize: 19,
@@ -69,32 +71,32 @@ class _NotaPageState extends State<NotaPage> {
                                   width: 150,
                                   height: 20,
                                   color: Colors.brown[100],
-                                  child: Text("Avaliação")),
-                              SizedBox(width: 5),
+                                  child: const Text("Avaliação")),
+                              const SizedBox(width: 5),
                               Container(
                                   alignment: Alignment.center,
                                   width: 45,
                                   height: 20,
                                   color: Colors.brown[100],
-                                  child: Text("Nota")),
-                              SizedBox(width: 5),
+                                  child: const Text("Nota")),
+                              const SizedBox(width: 5),
                               Container(
                                   alignment: Alignment.center,
                                   width: 45,
                                   height: 20,
                                   color: Colors.brown[100],
-                                  child: Text("Faltas")),
-                              SizedBox(width: 5),
+                                  child: const Text("Faltas")),
+                              const SizedBox(width: 5),
                               Container(
                                   alignment: Alignment.center,
                                   width: 45,
                                   height: 20,
                                   color: Colors.brown[100],
-                                  child: Text("Mais")),
+                                  child: const Text("Mais")),
                             ],
                           ),
                         ),
-                        SizedBox(height: 5),
+                        const SizedBox(height: 5),
                         Padding(
                           padding: const EdgeInsets.only(left: 5, right: 5),
                           child: Row(
@@ -104,26 +106,26 @@ class _NotaPageState extends State<NotaPage> {
                                   width: 150,
                                   height: 20,
                                   color: Colors.brown[50],
-                                  child: Text("1º semestre")),
-                              SizedBox(width: 5),
+                                  child:const Text("1º semestre")),
+                              const SizedBox(width: 5),
                               Container(
                                   alignment: Alignment.center,
                                   width: 45,
                                   height: 20,
                                   color: Colors.brown[50],
-                                  child: Text("12")),
-                              SizedBox(width: 5),
+                                  child:const Text("12")),
+                              const SizedBox(width: 5),
                               Container(
                                   alignment: Alignment.center,
                                   width: 45,
                                   height: 20,
                                   color: Colors.brown[50],
-                                  child: Text("1")),
+                                  child:const Text("1")),
                               Container(
                                 alignment: Alignment.center,
                                 width: 45,
                                 height: 20,
-                                child: Icon(
+                                child: const Icon(
                                   Icons.visibility,
                                   color: Colors.green,
                                 ),
@@ -140,26 +142,26 @@ class _NotaPageState extends State<NotaPage> {
                                   width: 150,
                                   height: 20,
                                   color: Colors.brown[50],
-                                  child: Text("2º semestre")),
-                              SizedBox(width: 5),
+                                  child:const Text("2º semestre")),
+                              const SizedBox(width: 5),
                               Container(
                                   alignment: Alignment.center,
                                   width: 45,
                                   height: 20,
                                   color: Colors.brown[50],
-                                  child: Text("12")),
-                              SizedBox(width: 5),
+                                  child:const Text("12")),
+                              const SizedBox(width: 5),
                               Container(
                                   alignment: Alignment.center,
                                   width: 45,
                                   height: 20,
                                   color: Colors.brown[50],
-                                  child: Text("5")),
+                                  child: const Text("5")),
                               Container(
                                 alignment: Alignment.center,
                                 width: 45,
                                 height: 20,
-                                child: Icon(
+                                child: const Icon(
                                   Icons.visibility,
                                   color: Colors.green,
                                 ),
@@ -176,26 +178,26 @@ class _NotaPageState extends State<NotaPage> {
                                   width: 150,
                                   height: 20,
                                   color: Colors.brown[50],
-                                  child: Text("----------")),
-                              SizedBox(width: 5),
+                                  child: const Text("----------")),
+                              const SizedBox(width: 5),
                               Container(
                                   alignment: Alignment.center,
                                   width: 45,
                                   height: 20,
                                   color: Colors.brown[50],
-                                  child: Text("--")),
-                              SizedBox(width: 5),
+                                  child: const Text("--")),
+                              const SizedBox(width: 5),
                               Container(
                                   alignment: Alignment.center,
                                   width: 45,
                                   height: 20,
                                   color: Colors.brown[50],
-                                  child: Text("--")),
+                                  child: const Text("--")),
                               Container(
                                   alignment: Alignment.center,
                                   width: 45,
                                   height: 20,
-                                  child: Text("--")),
+                                  child: const Text("--")),
                             ],
                           ),
                         ),
@@ -209,19 +211,19 @@ class _NotaPageState extends State<NotaPage> {
                                 height: 20,
                                 color: Colors.brown[50],
                               ),
-                              SizedBox(width: 5),
+                              const SizedBox(width: 5),
                               Container(
                                 alignment: Alignment.center,
                                 width: 95,
                                 height: 20,
-                                child: Text("Total de Faltas"),
+                                child: const Text("Total de Faltas"),
                                 color: Colors.brown[50],
                               ),
                               Container(
                                 alignment: Alignment.center,
                                 width: 50,
                                 height: 20,
-                                child: Text("6"),
+                                child:const   Text("6"),
                                 color: Colors.brown[50],
                               ),
                             ],
@@ -235,22 +237,22 @@ class _NotaPageState extends State<NotaPage> {
                                 alignment: Alignment.center,
                                 width: 150,
                                 height: 20,
-                                child: Text("Aprovado"),
+                                child: const  Text("Aprovado"),
                                 color: Colors.brown[50],
                               ),
-                              SizedBox(width: 5),
+                              const SizedBox(width: 5),
                               Container(
                                 alignment: Alignment.center,
                                 width: 50,
                                 height: 20,
-                                child: Text("Media"),
+                                child: const  Text("Media"),
                                 color: Colors.brown[50],
                               ),
                               Container(
                                 alignment: Alignment.center,
                                 width: 95,
                                 height: 20,
-                                child: Text("12"),
+                                child: const Text("12"),
                                 color: Colors.brown[50],
                               ),
                             ],
@@ -259,7 +261,7 @@ class _NotaPageState extends State<NotaPage> {
                       ],
                     ),
                   ),
-                  SizedBox(height: 10),
+                  const SizedBox(height: 10),
                 ],
               ),
             ),
