@@ -1,57 +1,26 @@
-import 'dart:convert';
-import 'i_pagamento_repository.dart';
+
+import 'package:dio/dio.dart';
+import 'package:sistema_escolar/repository/payment/i_pagamento_repository.dart';
 import 'package:sistema_escolar/shared/model/pagamento_model.dart';
-import 'package:http/http.dart' as http;
-import 'package:sistema_escolar/shared/model/user_model.dart';
+
 import 'package:sistema_escolar/utils/url_base.dart';
 
 class PagamentoRepository extends IPagamentoRepository {
+  final dio = Dio();
   @override
-  Future<List<Pagamento>> findAll() async {
-    var respomse = await http.get(UrlBase.pagamentos);
-    if (respomse.statusCode == 200) {
-      var data = json.decode(respomse.body);
-      List<Pagamento> pagamentos = PagamentoModel.fromJson(data).pagamentos;
-      return pagamentos;
-    } else {
-      throw Exception('Erro ao buscar pagamentos');
+  Future<List<Pagamento>?> findAll() async {
+    final response = await dio.get(UrlBase.pagamentos);
+    try {
+      return (response.data as List)
+          .map((data) => Pagamento.fromJson(data))
+          .toList();
+    } on DioError catch (dioError) {
+      return dioError.response!.data;
+    } catch (e) {
+      return [];
     }
   }
 
-  @override
-  Future<bool> delete(id) async {
-    var respomse = await http.delete(UrlBase.pagamentos);
-    if (respomse.statusCode == 200) {
-      return true;
-    } else {
-      return false;
-    }
-  }
 
-  @override
-  Future<List<Pagamento>> seachPagamento(String query) async {
-    var respomse = await http.get(UrlBase.pagamentos);
-    if (respomse.statusCode == 200) {
-      var data = json.decode(respomse.body);
-      List<Pagamento> pagamentos = PagamentoModel.fromJson(data).pagamentos;
-      return pagamentos;
-    } else {
-      throw Exception('Erro ao buscar pagamentos');
-    }
-  }
-
-  @override
-  Future<List<User>> findAllPagamento(String query) async {
-    var respomse = await http.get(UrlBase.pagamentos);
-    if (respomse.statusCode == 200) {
-      var data = json.decode(respomse.body);
-      List<User> user = User.fromJson(data) as List<User>;
-      return user;
-    } else {
-      throw Exception('Erro ao buscar pagamentos');
-    }
-  }
+ 
 }
-
-
-
