@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 
 class AppColors {
-  static Color p = Color.fromRGBO(90, 162, 255, 1);
-  static Color p1 = Color.fromRGBO(255, 255, 255, 1);
-  static Color secundy = Color.fromRGBO(111, 65, 239, 1);
-  static Color green = Color(0xFF1E8161);
-  static Color greenn = Color(0xFF8F8F);
+  static Color p = const Color.fromRGBO(90, 162, 255, 1);
+  static Color p1 =const Color.fromRGBO(255, 255, 255, 1);
+  static Color secundy =const Color.fromRGBO(111, 65, 239, 1);
+  static Color green =const Color(0xFF1E8161);
+  static Color greenn =const Color(0x00ff8f8f);
 }
