@@ -9,12 +9,12 @@ import 'package:sistema_escolar/pages/pagamento/pagamento_page.dart';
 import 'package:sistema_escolar/pages/perfil_page/perfil_page.dart';
 
 final Map<String, Widget Function(BuildContext)> appRoutes = {
-  'pagamento': (_) => DetalhePagamentoPage(),
-  'pagamentos': (_) => PagamentoPage(),
-  'falta': (_) => FaltaPage(),
-  'login': (_) => LoginPage(),
-  'dashboard': (_) => DashboardPage(),
-  'perfil': (_) => PerfilPage(),
-  'horario': (_) => HorarioPage(),
-  'notas': (_) => NotaPage(),
+  'pagamento': (_) =>const  DetalhePagamentoPage(),
+  'pagamentos': (_) =>const  PagamentoPage(),
+  'falta': (_) => const FaltaPage(),
+  'login': (_) => const LoginPage(),
+  'dashboard': (_) => const DashboardPage(),
+  'perfil': (_) => const PerfilPage(),
+  'horario': (_) => const HorarioPage(),
+  'notas': (_) => const NotaPage(),
 };
