@@ -1,69 +1,54 @@
-// To parse this JSON data, do
-//
-//     final pagamentoModel = pagamentoModelFromJson(jsonString);
-
 import 'dart:convert';
 
-PagamentoModel pagamentoModelFromJson(String str) => PagamentoModel.fromJson(json.decode(str));
-
-String pagamentoModelToJson(PagamentoModel data) => json.encode(data.toJson());
-
-class PagamentoModel {
-    PagamentoModel({
-        this.pagamentos,
-    });
-
-    List<Pagamento> pagamentos;
-
-    factory PagamentoModel.fromJson(Map<String, dynamic> json) => PagamentoModel(
-        pagamentos: List<Pagamento>.from(json["pagamentos"].map((x) => Pagamento.fromJson(x))),
-    );
-
-    Map<String, dynamic> toJson() => {
-        "pagamentos": List<dynamic>.from(pagamentos.map((x) => x.toJson())),
-    };
-}
-
 class Pagamento {
-    Pagamento({
-        this.id,
-        this.mes,
-        this.ano,
-        this.multa,
-        this.status,
-        this.valor,
-        this.createdAt,
-        this.updatedAt,
-    });
+  final int? id;
+  final String mes;
+  final String ano;
+  final String multa;
+  final String? status;
+  final int valor;
+  final DateTime? createdAt;
+  final DateTime? updatedAt;
+  Pagamento({
+    this.id,
+    required this.mes,
+    required this.ano,
+    required this.multa,
+    this.status,
+    required this.valor,
+    this.createdAt,
+    this.updatedAt,
+  });
 
-    int id;
-    String mes;
-    String ano;
-    String multa;
-    String status;
-    int valor;
-    DateTime createdAt;
-    DateTime updatedAt;
+ 
 
-    factory Pagamento.fromJson(Map<String, dynamic> json) => Pagamento(
-        id: json["id"],
-        mes: json["mes"],
-        ano: json["ano"],
-        multa: json["multa"],
-        status: json["status"],
-        valor: json["valor"],
-        createdAt: DateTime.parse(json["created_at"]),
-        updatedAt: DateTime.parse(json["updated_at"]),
-    );
-
-    Map<String, dynamic> toJson() => {
-        "id": id,
-        "mes": mes,
-        "ano": ano,
-        "multa": multa,
-        "status": status,
-        "valor": valor,
-        "created_at": createdAt.toIso8601String(),
-        "updated_at": updatedAt.toIso8601String(),
+  Map<String, dynamic> toMap() {
+    return {
+      'id': id,
+      'mes': mes,
+      'ano': ano,
+      'multa': multa,
+      'status': status,
+      'valor': valor,
+      'createdAt': createdAt?.millisecondsSinceEpoch,
+      'updatedAt': updatedAt?.millisecondsSinceEpoch,
     };
+  }
+
+  factory Pagamento.fromMap(Map<String, dynamic> map) {
+    return Pagamento(
+      id: map['id']?.toInt(),
+      mes: map['mes'] ?? '',
+      ano: map['ano'] ?? '',
+      multa: map['multa'] ?? '',
+      status: map['status'],
+      valor: map['valor']?.toInt() ?? 0,
+      createdAt: map['createdAt'] != null ? DateTime.fromMillisecondsSinceEpoch(map['createdAt']) : null,
+      updatedAt: map['updatedAt'] != null ? DateTime.fromMillisecondsSinceEpoch(map['updatedAt']) : null,
+    );
+  }
+
+  String toJson() => json.encode(toMap());
+
+  factory Pagamento.fromJson(String source) => Pagamento.fromMap(json.decode(source));
 }
