@@ -1,6 +1,6 @@
 class UrlBase {
-  static final pagamentos =
+  static const pagamentos =
       "http://192.168.43.204/2021/8-Agosto/escolar/public/api";
-  static final allUser =
+  static const allUser =
       "http://192.168.43.204/2021/8-Agosto/escolar/public/api";
 }
